% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat_plot.R
\name{confmat_plot}
\alias{confmat_plot}
\title{Confusion Matrix plot for Optimizing Probability Thresholds}
\usage{
confmat_plot(prob, ref, postarget, negtarget)
}
\arguments{
\item{prob}{A result probabilities from model}

\item{ref}{Actual target set}

\item{postarget}{Character positive levels}

\item{negtarget}{Character negative levels}
}
\value{
Confusion Matrix plot using \code{ggplotly}
}
\description{
Confusion Matrix plot for Optimizing Probability Thresholds
}
\examples{
dat <- tibble(
  "probability" = c(0.53, 0.64, 0.23, 0.1, 0.8)
  "churn_status" = as.factor(c("No","Yes", "No", "No", "Yes"))
)

dat \%>\% confat_plot(probability, churn_status, "Yes", "No")
}
